+++ src/main/java/kitchenpos/table/dto/OrderTableResponse.java
@@ -1,8 +1,8 @@
-package kitchenpos.dto;
+package kitchenpos.table.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 
 public class OrderTableResponse {
     private Long id;
